+++ git-config/src/file/git_config/mod.rs
@@ -4,14 +4,7 @@ mod resolve_includes;
 pub use from_env::functions::*;
 pub use resolve_includes::function::resolve_includes;
 
-use std::{borrow::Cow, collections::HashMap, convert::TryFrom};
-
-use crate::{
-    file::section::{MutableSection, SectionBody},
-    lookup,
-    parser::{ParsedSectionHeader, SectionHeaderName},
-    File,
-};
+use std::{borrow::Cow, collections::HashMap};
 
 /// The section ID is a monotonically increasing ID used to refer to sections.
 /// This value does not imply any ordering between sections, as new sections
@@ -37,445 +30,9 @@ pub(crate) enum LookupTreeNode<'a> {
     NonTerminal(HashMap<Cow<'a, str>, Vec<SectionId>>),
 }
 
-mod init;
-
-impl<'event> File<'event> {
-    /// Returns an interpreted value given a section, an optional subsection and
-    /// key.
-    ///
-    /// It's recommended to use one of the values in the [`values`] module as
-    /// the conversion is already implemented, but this function is flexible and
-    /// will accept any type that implements [`TryFrom<&[u8]>`][`TryFrom`].
-    ///
-    /// Consider [`Self::multi_value`] if you want to get all values of a
-    /// multivar instead.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use git_config::values::{Integer, Boolean};
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// let config = r#"
-    ///     [core]
-    ///         a = 10k
-    ///         c
-    /// "#;
-    /// let git_config = git_config::File::try_from(config)?;
-    /// // You can either use the turbofish to determine the type...
-    /// let a_value = git_config.value::<Integer>("core", None, "a")?;
-    /// // ... or explicitly declare the type to avoid the turbofish
-    /// let c_value: Boolean = git_config.value("core", None, "c")?;
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in the requested
-    /// section and subsection, if the section and subsection do not exist, or
-    /// if there was an issue converting the type into the requested variant.
-    ///
-    /// [`values`]: crate::values
-    /// [`TryFrom`]: std::convert::TryFrom
-    pub fn value<T: TryFrom<Cow<'event, [u8]>>>(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Result<T, lookup::Error<T::Error>> {
-        T::try_from(self.raw_value(section_name, subsection_name, key)?).map_err(lookup::Error::FailedConversion)
-    }
-
-    /// Like [`value()`][GitConfig::value()], but returning an `Option` if the value wasn't found.
-    pub fn try_value<T: TryFrom<Cow<'event, [u8]>>>(
-        &'event self,
-        section_name: &str,
-        subsection_name: Option<&str>,
-        key: &str,
-    ) -> Option<Result<T, T::Error>> {
-        self.raw_value(section_name, subsection_name, key).ok().map(T::try_from)
-    }
-
-    /// Returns all interpreted values given a section, an optional subsection
-    /// and key.
-    ///
-    /// It's recommended to use one of the values in the [`values`] module as
-    /// the conversion is already implemented, but this function is flexible and
-    /// will accept any type that implements [`TryFrom<&[u8]>`][`TryFrom`].
-    ///
-    /// Consider [`Self::value`] if you want to get a single value
-    /// (following last-one-wins resolution) instead.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use git_config::values::{Integer, Bytes, Boolean, TrueVariant};
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// let config = r#"
-    ///     [core]
-    ///         a = true
-    ///         c = g
-    ///     [core]
-    ///         a
-    ///         a = false
-    /// "#;
-    /// let git_config = git_config::File::try_from(config).unwrap();
-    /// // You can either use the turbofish to determine the type...
-    /// let a_value = git_config.multi_value::<Boolean>("core", None, "a")?;
-    /// assert_eq!(
-    ///     a_value,
-    ///     vec![
-    ///         Boolean::True(TrueVariant::Explicit(Cow::Borrowed("true"))),
-    ///         Boolean::True(TrueVariant::Implicit),
-    ///         Boolean::False(Cow::Borrowed("false")),
-    ///     ]
-    /// );
-    /// // ... or explicitly declare the type to avoid the turbofish
-    /// let c_value: Vec<Bytes> = git_config.multi_value("core", None, "c")?;
-    /// assert_eq!(c_value, vec![Bytes { value: Cow::Borrowed(b"g") }]);
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the key is not in the requested
-    /// section and subsection, if the section and subsection do not exist, or
-    /// if there was an issue converting the type into the requested variant.
-    ///
-    /// [`values`]: crate::values
-    /// [`TryFrom`]: std::convert::TryFrom
-    pub fn multi_value<'lookup, T: TryFrom<Cow<'event, [u8]>>>(
-        &'event self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-        key: &'lookup str,
-    ) -> Result<Vec<T>, lookup::Error<T::Error>> {
-        self.raw_multi_value(section_name, subsection_name, key)?
-            .into_iter()
-            .map(T::try_from)
-            .collect::<Result<Vec<_>, _>>()
-            .map_err(lookup::Error::FailedConversion)
-    }
-
-    /// Returns an immutable section reference.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the section and optional
-    /// subsection do not exist.
-    pub fn section<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-    ) -> Result<&SectionBody<'event>, lookup::existing::Error> {
-        let section_ids = self.section_ids_by_name_and_subname(section_name, subsection_name)?;
-        let id = section_ids.last().expect("BUG: Section lookup vec was empty");
-        Ok(self.sections.get(id).expect("BUG: Section did not have id from lookup"))
-    }
-
-    /// Returns an mutable section reference.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if the section and optional
-    /// subsection do not exist.
-    pub fn section_mut<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: Option<&'lookup str>,
-    ) -> Result<MutableSection<'_, 'event>, lookup::existing::Error> {
-        let section_ids = self.section_ids_by_name_and_subname(section_name, subsection_name)?;
-        let id = section_ids.last().expect("BUG: Section lookup vec was empty");
-        Ok(MutableSection::new(
-            self.sections
-                .get_mut(id)
-                .expect("BUG: Section did not have id from lookup"),
-        ))
-    }
-
-    /// Gets all sections that match the provided name, ignoring any subsections.
-    ///
-    /// # Examples
-    ///
-    /// Provided the following config:
-    ///
-    /// ```text
-    /// [core]
-    ///     a = b
-    /// [core ""]
-    ///     c = d
-    /// [core "apple"]
-    ///     e = f
-    /// ```
-    ///
-    /// Calling this method will yield all sections:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use git_config::values::{Integer, Boolean, TrueVariant};
-    /// # use std::borrow::Cow;
-    /// # use std::convert::TryFrom;
-    /// let config = r#"
-    ///     [core]
-    ///         a = b
-    ///     [core ""]
-    ///         c = d
-    ///     [core "apple"]
-    ///         e = f
-    /// "#;
-    /// let git_config = git_config::File::try_from(config).unwrap();
-    /// assert_eq!(git_config.sections_by_name("core").len(), 3);
-    /// ```
-    #[must_use]
-    pub fn sections_by_name<'lookup>(&self, section_name: &'lookup str) -> Vec<&SectionBody<'event>> {
-        self.section_ids_by_name(section_name)
-            .unwrap_or_default()
-            .into_iter()
-            .map(|id| {
-                self.sections
-                    .get(&id)
-                    .expect("section doesn't have id from from lookup")
-            })
-            .collect()
-    }
-
-    /// Get all sections that match the `section_name`, returning all matching section header along with their body.
-    ///
-    /// An empty `Vec` is returned if there is no section with `section_name`.
-    ///
-    /// # Example
-    ///
-    /// Provided the following config:
-    /// ```plain
-    /// [url "ssh://git@github.com/"]
-    ///     insteadOf = https://github.com/
-    /// [url "ssh://git@bitbucket.org"]
-    ///     insteadOf = https://bitbucket.org/
-    /// ```
-    /// Calling this method will yield all section bodies and their header:
-    ///
-    /// ```rust
-    /// use git_config::File;
-    /// use git_config::parser::Key;
-    /// use std::borrow::Cow;
-    /// use std::convert::TryFrom;
-    /// use nom::AsBytes;
-    ///
-    /// let input = r#"
-    /// [url "ssh://git@github.com/"]
-    ///    insteadOf = https://github.com/
-    /// [url "ssh://git@bitbucket.org"]
-    ///    insteadOf = https://bitbucket.org/
-    /// "#;
-    /// let config = git_config::File::try_from(input).unwrap();
-    /// let url = config.sections_by_name_with_header("url");
-    /// assert_eq!(url.len(), 2);
-    ///
-    /// for (i, (header, body)) in url.iter().enumerate() {
-    ///     let url = header.subsection_name.as_ref();
-    ///     let instead_of = body.value(&Key::from("insteadOf"));
-    ///
-    ///     // todo(unstable-order): the order is not always the same, so `i` cannot be used here
-    ///     if instead_of.as_ref().unwrap().as_ref().as_bytes().eq("https://github.com/".as_bytes()) {
-    ///         assert_eq!(instead_of.unwrap().as_ref(), "https://github.com/".as_bytes());
-    ///         assert_eq!(url.unwrap().as_ref(), "ssh://git@github.com/");
-    ///     } else {
-    ///         assert_eq!(instead_of.unwrap().as_ref(), "https://bitbucket.org/".as_bytes());
-    ///         assert_eq!(url.unwrap().as_ref(), "ssh://git@bitbucket.org");
-    ///     }
-    /// }
-    /// ```
-    pub fn sections_by_name_with_header<'lookup>(
-        &self,
-        section_name: &'lookup str,
-    ) -> Vec<(&ParsedSectionHeader<'event>, &SectionBody<'event>)> {
-        self.section_ids_by_name(section_name)
-            .unwrap_or_default()
-            .into_iter()
-            .map(|id| {
-                (
-                    self.section_headers
-                        .get(&id)
-                        .expect("section doesn't have a section header??"),
-                    self.sections
-                        .get(&id)
-                        .expect("section doesn't have id from from lookup"),
-                )
-            })
-            .collect()
-    }
-
-    /// Adds a new section to config. If a subsection name was provided, then
-    /// the generated header will use the modern subsection syntax. Returns a
-    /// reference to the new section for immediate editing.
-    ///
-    /// # Examples
-    ///
-    /// Creating a new empty section:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::convert::TryFrom;
-    /// let mut git_config = git_config::File::new();
-    /// let _section = git_config.new_section("hello", Some("world".into()));
-    /// assert_eq!(git_config.to_string(), "[hello \"world\"]\n");
-    /// ```
-    ///
-    /// Creating a new empty section and adding values to it:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::convert::TryFrom;
-    /// let mut git_config = git_config::File::new();
-    /// let mut section = git_config.new_section("hello", Some("world".into()));
-    /// section.push("a".into(), "b".as_bytes().into());
-    /// assert_eq!(git_config.to_string(), "[hello \"world\"]\n  a=b\n");
-    /// let _section = git_config.new_section("core", None);
-    /// assert_eq!(git_config.to_string(), "[hello \"world\"]\n  a=b\n[core]\n");
-    /// ```
-    pub fn new_section(
-        &mut self,
-        section_name: impl Into<Cow<'event, str>>,
-        subsection_name: impl Into<Option<Cow<'event, str>>>,
-    ) -> MutableSection<'_, 'event> {
-        let mut section = self.push_section(section_name, subsection_name, SectionBody::new());
-        section.push_newline();
-        section
-    }
-
-    /// Removes the section, returning the events it had, if any. If multiple
-    /// sections have the same name, then the last one is returned. Note that
-    /// later sections with the same name have precedent over earlier ones.
-    ///
-    /// # Examples
-    ///
-    /// Creating and removing a section:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::convert::TryFrom;
-    /// let mut git_config = git_config::File::try_from(
-    /// r#"[hello "world"]
-    ///     some-value = 4
-    /// "#).unwrap();
-    ///
-    /// let events = git_config.remove_section("hello", Some("world".into()));
-    /// assert_eq!(git_config.to_string(), "");
-    /// ```
-    ///
-    /// Precedence example for removing sections with the same name:
-    ///
-    /// ```
-    /// # use git_config::File;
-    /// # use std::convert::TryFrom;
-    /// let mut git_config = git_config::File::try_from(
-    /// r#"[hello "world"]
-    ///     some-value = 4
-    /// [hello "world"]
-    ///     some-value = 5
-    /// "#).unwrap();
-    ///
-    /// let events = git_config.remove_section("hello", Some("world".into()));
-    /// assert_eq!(git_config.to_string(), "[hello \"world\"]\n    some-value = 4\n");
-    /// ```
-    pub fn remove_section<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: impl Into<Option<&'lookup str>>,
-    ) -> Option<SectionBody> {
-        let id = self
-            .section_ids_by_name_and_subname(section_name, subsection_name.into())
-            .ok()?
-            .pop()?;
-        self.section_order.remove(
-            self.section_order
-                .iter()
-                .position(|v| *v == id)
-                .expect("Section order does not contain section that we were trying to remove"),
-        );
-        self.sections.remove(&id)
-    }
-
-    /// Adds the provided section to the config, returning a mutable reference
-    /// to it.
-    pub fn push_section(
-        &mut self,
-        section_name: impl Into<Cow<'event, str>>,
-        subsection_name: impl Into<Option<Cow<'event, str>>>,
-        section: SectionBody<'event>,
-    ) -> MutableSection<'_, 'event> {
-        let subsection_name = subsection_name.into();
-        if subsection_name.is_some() {
-            self.push_section_internal(
-                ParsedSectionHeader {
-                    name: SectionHeaderName(section_name.into()),
-                    separator: Some(" ".into()),
-                    subsection_name,
-                },
-                section,
-            )
-        } else {
-            self.push_section_internal(
-                ParsedSectionHeader {
-                    name: SectionHeaderName(section_name.into()),
-                    separator: None,
-                    subsection_name: None,
-                },
-                section,
-            )
-        }
-    }
-
-    /// Renames a section, modifying the last matching section.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the lookup doesn't exist
-    pub fn rename_section<'lookup>(
-        &mut self,
-        section_name: &'lookup str,
-        subsection_name: impl Into<Option<&'lookup str>>,
-        new_section_name: impl Into<SectionHeaderName<'event>>,
-        new_subsection_name: impl Into<Option<Cow<'event, str>>>,
-    ) -> Result<(), lookup::existing::Error> {
-        let id = self.section_ids_by_name_and_subname(section_name, subsection_name.into())?;
-        let id = id
-            .last()
-            .expect("list of sections were empty, which violates invariant");
-        let header = self
-            .section_headers
-            .get_mut(id)
-            .expect("sections does not have section id from section ids");
-        header.name = new_section_name.into();
-        header.subsection_name = new_subsection_name.into();
-
-        Ok(())
-    }
-
-    /// Returns the number of values in the config, no matter in which section.
-    ///
-    /// For example, a config with multiple empty sections will return 0.
-    /// This ignores any comments.
-    #[must_use]
-    pub fn num_values(&self) -> usize {
-        self.sections
-            .values()
-            .fold(0, |acc, section| acc + section.num_values())
-    }
-
-    /// Returns if there are no entries in the config. This will return true
-    /// if there are only empty sections or comments.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.sections.values().all(SectionBody::is_empty)
-    }
-}
-
 mod access;
 mod impls;
+mod init;
 mod utils;
 
 #[cfg(test)]
